edited entities in core microservice
