+++ core/src/main/java/com/sahan/core/Entities/Market.java
@@ -1,4 +1,4 @@
-package com.sahan.market.Entities;
+package com.sahan.core.Entities;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
