+++ core/src/main/java/com/sahan/core/Entities/Order.java
@@ -1,0 +1,4 @@
+package com.sahan.core.Entities;
+
+public class Order {
+}
