--- core/src/main/java/com/sahan/market/Entities/OrderReturn.java
@@ -1,4 +1,0 @@
-package com.sahan.market.Entities;
-
-public class OrderReturn {
-}
